+++ src/tableplus/core.py
@@ -1,11 +1,8 @@
-from argparse import Namespace
 from pathlib import Path
 from urllib.parse import quote
 from urllib.parse import urlencode
 
-import environ
-
-# from urllib.parse import quote_plus
+from dotenv import dotenv_values
 
 
 def build_db_url(  # noqa: PLR0913
@@ -30,17 +27,9 @@ def build_db_url(  # noqa: PLR0913
     )
 
 
-def get_env(env_path=None) -> environ.Env:
-    """Get the environment variables."""
-    env = environ.Env()
-    if env_path and Path(env_path).exists():
-        env.read_env(env_path)
-    return env
-
-
 def get_local_db_conn_str(env_path=None) -> str:
     """Get the local DB connection string."""
-    env = get_env(env_path)
+    env = dotenv_values(Path(env_path))
     params = {
         "statusColor": "F8F8F8",
         "env": "local",
@@ -48,16 +37,16 @@ def get_local_db_conn_str(env_path=None) -> str:
         "lazyload": "true",
     }
 
-    db_user = env("POSTGRES_USER")
-    db_pass = env("POSTGRES_PASSWORD")
-    db_name = env("POSTGRES_DB")
+    db_user = env["POSTGRES_USER"]
+    db_pass = env["POSTGRES_PASSWORD"]
+    db_name = env["POSTGRES_DB"]
 
     return build_db_url(db_user, db_pass, db_name, params=params)
 
 
 def get_prod_db_conn_str(env_path, ssh_user: str, ssh_host: str) -> str:
     """Get the production DB connection string."""
-    env = get_env(env_path)
+    env = dotenv_values(Path(env_path))
     params = {
         "statusColor": "FFD7D4",
         "env": "production",
@@ -70,9 +59,9 @@ def get_prod_db_conn_str(env_path, ssh_user: str, ssh_host: str) -> str:
         "lazyload": "true",
     }
 
-    db_user = env("POSTGRES_USER")
-    db_pass = env("POSTGRES_PASSWORD")
-    db_name = env("POSTGRES_DB")
+    db_user = env["POSTGRES_USER"]
+    db_pass = env["POSTGRES_PASSWORD"]
+    db_name = env["POSTGRES_DB"]
 
     return build_db_url(
         db_user,
@@ -83,25 +72,3 @@ def get_prod_db_conn_str(env_path, ssh_user: str, ssh_host: str) -> str:
         ssh_user=ssh_user,
         ssh_host=ssh_host,
     )
-
-
-def _output_db_urls(local_db_url: str, prod_db_url: str) -> None:
-    """Output the DB URLs."""
-    print("=> TablePlus: Right click > New > Connection from URL...")  # noqa: T201
-    print("LOCAL:", local_db_url)  # noqa: T201
-    print("PROD:", prod_db_url)  # noqa: T201
-
-
-def run(args: Namespace) -> None:
-    """Main entry point for the CLI."""
-    project_path = Path(args.path).expanduser()
-    local_env_path = str(project_path / ".envs/.local/.postgres")
-    prod_env_path = str(project_path / ".envs/.production/.postgres")
-
-    # local_env = get_env(local_env_path)
-    # prod_env = get_env(prod_env_path)
-
-    local_db_url = get_local_db_conn_str(local_env_path)
-    prod_db_url = get_prod_db_conn_str(prod_env_path, args.ssh_user, args.ssh_host)
-
-    _output_db_urls(local_db_url, prod_db_url)