+++ HISTORY.md
@@ -1,5 +1,13 @@
 # History
+
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
-## 0.0.1 (2024-04-09)
-* First release
+## 0.1.1 (2024-04-23)
+
+- Moved from `argparse` to `click`
+- Moved from `python-environ` to `python-dotenv`
+- Fixed bug with production URL
+
+## 0.1.0 (2024-04-09)
+
+- First release